+++ Packages/com.zebugger.zebug/Runtime/DebugGui/ZebugDebugGui.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using JetBrains.Annotations;
 using TMPro;
 using UnityEngine;
+using UnityEngine.Events;
 using UnityEngine.InputSystem;
 using UnityEngine.UI;
 using ZebugProject;
@@ -11,6 +12,7 @@ public class ZebugDebugGui : MonoBehaviour
     [SerializeField] private Button _tabButtonPrefab;
     [SerializeField] private GameObject _tabPanelPrefab;
     [SerializeField] private ZebugTabVarElement _tabPanelVarElementPrefab;
+    [SerializeField] private ZebugTabButtonElement _tabPanelButtonElementPrefab;
     
     [SerializeField] private GameObject _tabButtonPanelRoot;
     [SerializeField] private GameObject _tabPanelRoot;
@@ -87,36 +89,74 @@ public class ZebugDebugGui : MonoBehaviour
         public Button button;
         public GameObject varElementPanel;
         public List<ZebugTabVarElement> vars = new();
+        public List<ZebugTabButtonElement> buttons = new();
     }
     
     
     private Dictionary<IChannel, Tab> _channelTabs = new();
 
+    private void GetTabForChannel(IChannel channel, out Tab tab)
+    {
+        if (!_channelTabs.TryGetValue(channel, out tab))
+        {
+            var tabButton = Instantiate(_tabButtonPrefab, _tabButtonPanelRoot.transform);
+            tabButton.GetComponentInChildren<TMP_Text>().text = channel.Name();
+            tabButton.onClick.AddListener(() =>
+            {
+                SetActiveTab(channel);
+            });
+                
+            var tabPanel = Instantiate(_tabPanelPrefab, _tabPanelRoot.transform);
+            tabPanel.SetActive(false);
+                
+            tab = new Tab()
+            {
+                button = tabButton,
+                varElementPanel = tabPanel
+            };
+                
+            _channelTabs.Add(channel, tab);
+        }
+    }
+    
     private void CheckInit()
     {
+        foreach (var (channel, windowButtons) in Zebug.s_ChannelWindowButtons)
+        {
+            // find tab for channel
+            GetTabForChannel(channel, out var tab);
+            
+            if (!tab.varElementPanel.activeSelf)
+            {
+                continue;
+            }
+
+            int varCount = windowButtons.Count;
+            for (int i = tab.buttons.Count - 1; i >= varCount; i--)
+            {
+                Destroy(tab.buttons[i].gameObject);
+                tab.buttons.RemoveAt(i);
+            }
+            for (int i = tab.vars.Count; i < varCount; i++)
+            {
+                var tabVar = Instantiate(_tabPanelButtonElementPrefab, tab.varElementPanel.transform);
+                tab.buttons.Add(tabVar);
+            }
+            
+            int varIdx = 0;
+            foreach (var (varName, callback) in windowButtons)
+            {
+                var tabVar = tab.buttons[varIdx++];
+                tabVar.varName.text = varName;
+                tabVar.varButton.onClick.RemoveAllListeners();
+                tabVar.varButton.onClick.AddListener(()=>callback?.Invoke());
+            }
+        }
+        
         foreach (var (channel, windowVar) in Zebug.s_ChannelWindowVariables)
         {
             // find tab for channel
-            if (!_channelTabs.TryGetValue(channel, out var tab))
-            {
-                var tabButton = Instantiate(_tabButtonPrefab, _tabButtonPanelRoot.transform);
-                tabButton.GetComponentInChildren<TMP_Text>().text = channel.Name();
-                tabButton.onClick.AddListener(() =>
-                {
-                    SetActiveTab(channel);
-                });
-                
-                var tabPanel = Instantiate(_tabPanelPrefab, _tabPanelRoot.transform);
-                tabPanel.SetActive(false);
-                
-                tab = new Tab()
-                {
-                    button = tabButton,
-                    varElementPanel = tabPanel
-                };
-                
-                _channelTabs.Add(channel, tab);
-            }
+            GetTabForChannel(channel, out var tab);
             
             if (!tab.varElementPanel.activeSelf)
             {
@@ -147,30 +187,22 @@ public class ZebugDebugGui : MonoBehaviour
 
     private void SetActiveTab([CanBeNull]IChannel targetChannel)
     {
-        foreach (var (channel, _) in Zebug.s_ChannelWindowVariables)
+        foreach ((IChannel channel, Tab tab) in _channelTabs)
         {
-            if (channel == null)
-            {
-                continue;
-            }
-            
             bool enable = targetChannel == channel;
             
-            if (_channelTabs.TryGetValue(channel, out var tab))
+            if (enable)
             {
-                if (enable)
+                bool already = tab.varElementPanel.activeSelf;
+                if (already)
                 {
-                    bool already = tab.varElementPanel.activeSelf;
-                    if (already)
-                    {
-                        //  --- Toggle off.
-                        enable = false;
-                    }
+                    //  --- Toggle off.
+                    enable = false;
                 }
-                
-                tab.button.GetComponent<Image>().color = enable ? Color.white : new Color(0.72f, 0.72f, 0.72f);
-                tab.varElementPanel.SetActive(enable);
             }
+                
+            tab.button.GetComponent<Image>().color = enable ? Color.white : new Color(0.72f, 0.72f, 0.72f);
+            tab.varElementPanel.SetActive(enable);
         }
     }
 }